+++ apps/api/src/modules/role/index.ts
@@ -1,0 +1,5 @@
+export * from './role.module';
+export * from './role.service';
+export * from './role.controller';
+export * from './dto/create-role.dto';
+export * from './dto/update-role.dto';